diabetes-analytics-backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
pydantic
pandas
numpy
scikit-learn
PyMySQL
python-dotenv
